requirements.txt
# Core data processing
polars>=0.20.0
pyarrow>=14.0.0

# MSSQL connectivity
pyodbc>=5.0.0

# Encryption
cryptography>=41.0.0

# Logging
python-json-logger>=2.0.7

# Configuration
pyyaml>=6.0.1

# Date utilities
python-dateutil>=2.8.2
